# Conflicts:
#	.idea/workspace.xml

+++ src/algorithm/Algorithm.java
@@ -10,7 +10,7 @@ import java.util.Random;
  * Created by seal on 3/11/15.
  */
 public class Algorithm {
-    protected int[][] flowMatrix;
+    protected double[][] flowMatrix;
     protected boolean[][] connectionMatrix;
     protected int[] markMatrix;     //marking where a node is demand or supply; 0 for supply and 1 for demand.
     protected Random random;
@@ -19,7 +19,7 @@ public class Algorithm {
      * Initialize all member variable;
      */
     protected void initVariable(int size) {
-        flowMatrix = new int[size][size];
+        flowMatrix = new double[size][size];
         connectionMatrix = new boolean[size][size];
         markMatrix = new int[size];
         random = new Random();
@@ -49,7 +49,7 @@ public class Algorithm {
      *  method check the return value of the node's getSupplyOrDemand. If positive then it is
      *  supply node otherwise demand node. This method return 0 for the supply and 1 for the demand.
      */
-    protected int isDemandOrSupply(int val) {
+    protected int isDemandOrSupply(double val) {
         return (val > 0) ? 0 : 1;
     }
 
@@ -65,7 +65,7 @@ public class Algorithm {
                 if (connectionMatrix[i][j]) {
                     int index = indexOf(listOfEdge, j);
                     if (listOfEdge.get(index).getCapacity() <= 0) continue;
-                    int randomValue = random.nextInt(listOfEdge.get(index).getCapacity());
+                    double randomValue = (double)random.nextInt(listOfEdge.get(index).getCapacity());
                     flowMatrix[i][j] = -randomValue;
                     flowMatrix[j][i] = randomValue;
                 }
@@ -90,12 +90,12 @@ public class Algorithm {
      */
     protected boolean checkValidity(List<Node> graph) {
         int i = 0;
-        for(int[] row : flowMatrix) {
+        for(double[] row : flowMatrix) {
             if (markMatrix[i] == 0) {
-                int supply = graph.get(i).getSupply();
+                double supply = graph.get(i).getSupply();
                 if (!checkValiditySupply(row, supply)) return false;
             } else {
-                int demand = graph.get(i).getDemand();
+                double demand = graph.get(i).getDemand();
                 if (!checkValidityDemand(row, demand)) return false;
             }
             i++;
@@ -107,12 +107,12 @@ public class Algorithm {
      * This method check for the supply condition. That is total incoming and it self generated
      * supply should be greater then is outgoing.
      */
-    protected boolean checkValiditySupply(int[] row, int selfGeneration) {
-        int total = 0;
+    protected boolean checkValiditySupply(double[] row, double selfGeneration) {
+        double total = 0.0;
 
-        for (int i : row) total += i;
+        for (double i : row) total += i;
 
-        return (total + selfGeneration) >= 0;
+        return (total + selfGeneration) >= 0.0;
     }
 
     /**
@@ -120,10 +120,10 @@ public class Algorithm {
      *  smaller than the total incoming. Therefor the total incoming and outgoing summation
      *  is greater than 0.
      */
-    protected boolean checkValidityDemand(int[] row, int demand) {
-        int total = 0;
+    protected boolean checkValidityDemand(double[] row, double demand) {
+        double total = 0.0;
 
-        for (int i : row) total += i;
+        for (double i : row) total += i;
 
         return total >= 0;
     }
@@ -132,13 +132,13 @@ public class Algorithm {
      * Calculate the total load shedding of the demand node and return the value in negative form because the demand
      * in the negative from. We assume that there is no load shedding in the supply node.
      */
-    protected int totalLoadShedding(List<Node> graph) {
-        int totalLoadShedding = 0;
+    protected double totalLoadShedding(List<Node> graph) {
+        double totalLoadShedding = 0.0;
         for (int i = 0; i < flowMatrix.length; i++) {
             if (markMatrix[i] == 1) {
-                int total = 0;
+                double total = 0.0;
 
-                for (int j : flowMatrix[i]) total += j;
+                for (double j : flowMatrix[i]) total += j;
 
                 totalLoadShedding += (total + graph.get(i).getDemand());
             }
